Consolidate rules imports
