+++ src/sqlfluff/utils/reflow/elements.py
@@ -16,8 +16,7 @@ from sqlfluff.core.parser.segments import (
     TemplateSegment,
     WhitespaceSegment,
 )
-from sqlfluff.core.rules import LintFix
-from sqlfluff.core.rules.base import LintResult
+from sqlfluff.core.rules import LintFix, LintResult
 from sqlfluff.utils.reflow.config import ReflowConfig
 from sqlfluff.utils.reflow.depthmap import DepthInfo
