+++ src/sqlfluff/utils/reflow/respace.py
@@ -10,8 +10,7 @@ from sqlfluff.core.parser import (
     RawSegment,
     WhitespaceSegment,
 )
-from sqlfluff.core.rules import LintFix
-from sqlfluff.core.rules.base import LintResult
+from sqlfluff.core.rules import LintFix, LintResult
 from sqlfluff.utils.reflow.helpers import pretty_segment_name
 
 if TYPE_CHECKING:  # pragma: no cover
